chore: Support applyTheme from window API
