+++ packages/threat-composer/src/customTypes/windowAPI.ts
@@ -13,18 +13,10 @@
   See the License for the specific language governing permissions and
   limitations under the License.
  ******************************************************************************************************************** */
-export * from './assumptions';
-export * from './mitigations';
-export * from './threats';
-export * from './threatFieldTypes';
-export * from './workspaces';
-export * from './entities';
-export * from './composerMode';
-export * from './application';
-export * from './architecture';
-export * from './dataflow';
-export * from './dataExchange';
-export * from './events';
-export * from './components';
-export * from './referencePacks';
-export * from './windowAPI';
+import { ThreatComposerNamespace } from './dataExchange';
+
+declare global {
+  interface Window {
+    threatcomposer: ThreatComposerNamespace;
+  }
+}
